Move UnprotectedRootAction class under lib directory
